[codewhale] only remove volume when forcing build ;

+++ assistant/codewhale/0.8.47/build.sh
@@ -1,4 +1,6 @@
 #!/bin/sh
-docker volume rm -f codewhale-home
+if [ "${1:-}" = "-f" ]; then
+    docker volume rm -f codewhale-home
+fi
 docker rmi local/codewhale:v0.8.47-hardened 
 docker build -t local/codewhale:v0.8.47-hardened docker
